requirements.txt
aiohttp==3.7.4.post0
beautifulsoup4==4.9.1
discord.py==2.2.2
matplotlib==3.5.1
numpy==1.24.4
python-dotenv==1.0.1
python_dateutil==2.8.2
python_Levenshtein==0.25.0
Requests==2.31.0
scipy==1.9.1
